Renamed files and moved validator.rs to validators/mod.rs
